backend/requirements_dev.txt
-r requirements.txt
ipython
pytest-django==4.7.0
